refactor(to2txt): additional cleanup in parser lib
