+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "to2txt"
-version = "0.1.1"
+version = "1.0.0-beta.1"
 authors = ["Zachary Golba <zachary.golba@postlight.com>"]
 edition = "2021"
 license = "MIT OR Apache-2.0"
