+++ src/parser.rs
@@ -1,5 +1,5 @@
 use std::borrow::Cow;
-use std::str::FromStr;
+use std::str::{CharIndices, FromStr};
 
 use chrono::NaiveDate;
 use nom::bytes::complete::{is_not, tag, take};
@@ -9,15 +9,52 @@ use nom::sequence::{delimited, preceded, terminated};
 use nom::{IResult, Parser};
 use nom_locate::position;
 
-use crate::todo::{Located, Priority, Span, Todo};
+use crate::todotxt::{Located, Priority, Span, Tag, Todo};
 
 pub type Error<'a> = nom::error::Error<Input<'a>>;
 pub type Input<'a> = nom_locate::LocatedSpan<&'a str>;
+
+struct SplitWhitespace<'a> {
+    iter: CharIndices<'a>,
+    len: usize,
+}
 
 /// Parse a todo list from the provided `&str`.
 ///
 pub fn from_str(value: &str) -> impl Iterator<Item = Todo> {
     iterator(value.into(), delimited(eol, todo(), eol))
+}
+
+pub fn tags(at: Span, description: &str) -> impl Iterator<Item = Tag> {
+    let line = at.line();
+    let offset = at.start();
+
+    SplitWhitespace::new(description).filter_map(move |(start, end)| {
+        let token = &description[start..end];
+        let span = Span::new(line, (start + offset, end + offset));
+
+        if token.starts_with('@') && token.len() > 1 {
+            return Some(Tag::Context(Located {
+                data: &token[1..],
+                span,
+            }));
+        }
+
+        if token.starts_with('+') && token.len() > 1 {
+            return Some(Tag::Project(Located {
+                data: &token[1..],
+                span,
+            }));
+        }
+
+        token.split_once(':').and_then(|pair| {
+            if !pair.0.is_empty() && !pair.1.is_empty() {
+                Some(Tag::Named(Located { data: pair, span }))
+            } else {
+                None
+            }
+        })
+    })
 }
 
 /// Returns a parser that parses a single task on a todo list.
@@ -110,4 +147,28 @@ fn ymd<'a>() -> impl Parser<Input<'a>, Output = Located<NaiveDate>, Error = Erro
             })
         },
     )
+}
+
+impl<'a> SplitWhitespace<'a> {
+    fn new(value: &'a str) -> Self {
+        Self {
+            iter: value.char_indices(),
+            len: value.len(),
+        }
+    }
+}
+
+impl<'a> Iterator for SplitWhitespace<'a> {
+    type Item = (usize, usize);
+
+    fn next(&mut self) -> Option<Self::Item> {
+        let iter = &mut self.iter;
+
+        let start = iter.find_map(|(i, c)| if c.is_whitespace() { None } else { Some(i) })?;
+        let end = iter
+            .find_map(|(i, c)| if c.is_whitespace() { Some(i) } else { None })
+            .unwrap_or(self.len);
+
+        Some((start, end))
+    }
 }
